+++ back/src/main/java/com/bbt/back/model/CollectPic.java
@@ -9,32 +9,32 @@ import java.util.HashMap;
  * @Date: 2019/10/28 0028 16:26
  */
 public class CollectPic implements Serializable {
-    Integer collectNum;
+    Long collectNum;
     Integer rank;
-    HashMap<String,Integer> map;
+    HashMap<String,Long> map;
 
     public CollectPic() {
     }
 
-    public CollectPic(Integer collectNum, Integer rank, HashMap<String, Integer> map) {
+    public CollectPic(Long collectNum, Integer rank, HashMap<String, Long> map) {
         this.collectNum = collectNum;
         this.rank = rank;
         this.map = map;
     }
 
-    public HashMap<String, Integer> getMap() {
+    public HashMap<String, Long> getMap() {
         return map;
     }
 
-    public void setMap(HashMap<String, Integer> map) {
+    public void setMap(HashMap<String, Long> map) {
         this.map = map;
     }
 
-    public Integer getCollectNum() {
+    public Long getCollectNum() {
         return collectNum;
     }
 
-    public void setCollectNum(Integer collectNum) {
+    public void setCollectNum(Long collectNum) {
         this.collectNum = collectNum;
     }
